--- src/app/ui/header.tsx
@@ -1,11 +1,0 @@
-import '@/styles/header.css'
-
-export function Header() {
-  return (
-    <div className="header">
-      <h1>
-        <img src="brewfinder_logo.png" /> Brew Finder
-      </h1>
-    </div>
-  )
-}
