--- src/styles/footer.css
@@ -1,7 +1,0 @@
-.footer {
-    background-color: grey;
-    padding-bottom: 250px;
-    margin-bottom: -250px;
-    position: relative;
-    z-index: 1000;
-}
